Development DSKY

+++ src/gauges/base/AbstractGauge.cs
@@ -135,7 +135,7 @@ namespace Nereid
          // draw a rectagle at (x,y) of with w, height h of color c:
          private readonly Texture2D RECT_TEXTURE = new Texture2D(1, 1);
          private Rect RECT_BOUNDS = new Rect();
-         protected void DrawRectagle(int x, int y, int w, int h, Color c)
+         protected void DrawRectagle(float x, float y, float w, float h, Color c)
          {
             RECT_TEXTURE.SetPixel(0, 0, c);
             RECT_TEXTURE.Apply();
